move proxy systems into separate module
